mrc-4763: Add endpoints to serve emulator data.
If a `--emulator=PATH` argument is passed to the mintr main, two new
endpoints are provided, one to serve a JSON configuration and the other
to serve the ML model used by the emulator.

The endpoints are pretty agnostic to the actual format and contents of
the data. This allows a lot of flexibility while we iterate over the
client-side implementation of the emulator. In the future we may narrow
this down, eg. by specifying a more restrictive JSON Schema for the
configuration endpoint.

mintr can still function without any emulator data: if the `--emulator`
argument is missing then the endpoints are not included, and calling
them returns a 404.